refactor(api): add year workspace
