+++ apps/api/src/index.ts
@@ -11,7 +11,7 @@ import subjectsRoutes from "@/routes/subjects";
 import usersRoutes from "@/routes/users";
 import averagesRoute from "@/routes/averages";
 import feedbackRoute from "@/routes/feedback";
-import cardsRoute from "@/routes/cards";
+// import cardsRoute from "@/routes/cards";
 import { Hono } from "hono";
 import { cors } from "hono/cors";
 import { logger } from "hono/logger";
@@ -62,7 +62,7 @@ app.route("/averages", averagesRoute);
 
 app.route("/feedback", feedbackRoute);
 
-app.route("/cards", cardsRoute);
+// app.route("/cards", cardsRoute);
 
 app.all("/uploadthing", (ctx) => uploadHandlers(ctx.req.raw));
