+++ apps/api/src/lib/nanoid.ts
@@ -11,11 +11,12 @@ per - period
 ca - custom average
 ct - card template
 cl - card layout
+y - year
 */
 
-type IdPrefix = "u" | "acc" | "ses" | "ver" | "sub" | "gra" | "per" | "ca" | "ct" | "cl";
+type IdPrefix = "u" | "acc" | "ses" | "ver" | "sub" | "gra" | "per" | "ca" | "ct" | "cl" | "y";
 
-const ALPHABET = "0123456789abcdefghijklmnopqrstuvwxyz";
+const ALPHABET = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
 
 const nanoid = customAlphabet(ALPHABET, 12);
