+++ apps/api/src/routes/grades.ts
@@ -16,138 +16,22 @@ const app = new Hono<{
   };
 }>();
 
-/**
- * Create a new grade
- */
-const createGradeSchema = z.object({
-  name: z.string().min(1).max(64),
-  outOf: z
-    .number()
-    .min(0)
-    .max(1000 * 10)
-    .transform((f) => Math.round(f * 100)),
-  value: z
-    .number()
-    .min(0)
-    .max(1000 * 10)
-    .transform((f) => Math.round(f * 100)),
-  coefficient: z
-    .number()
-    .min(0)
-    .max(1000 * 10)
-    .transform((f) => Math.round(f * 100)),
-  passedAt: z.coerce.date().refine((date) => date <= new Date(), {
-      message: "Date cannot be in the future",
-    }).optional().default(new Date()),
-  subjectId: z.string().min(1).max(64),
-  periodId: z.string().min(1).max(64).nullable(),
-});
-
-app.post("/", zValidator("json", createGradeSchema), async (c) => {
-  const session = c.get("session");
-
-  if (!session) throw new HTTPException(401);
-
-  // If email isnt verified
-  if (!session.user.emailVerified) {
-    return c.json(
-      {
-        code: "EMAIL_NOT_VERIFIED",
-        message: "Email verification is required",
-      },
-      403
-    );
-  }
-
-  const { name, outOf, value, coefficient, passedAt, subjectId, periodId } =
-    c.req.valid("json");
-
-  const subject = await db.query.subjects.findFirst({
-    where: eq(subjects.id, subjectId),
+async function getMarkById(markId: string) {
+  const mark = await db.query.grades.findFirst({
+    where: (marks, { eq }) => eq(marks.id, markId),
   });
+  return mark;
+}
 
-  if (!subject) throw new HTTPException(404);
-  if (subject.userId !== session.user.id) throw new HTTPException(403);
-
-
-    // Check if the period exists and belongs to the user
-  if (periodId) {
-    const period = await db.query.periods.findFirst({
-      where: eq(periods.id, periodId),
-    });
-
-    if (!period) throw new HTTPException(404);
-    if (period.userId !== session.user.id) throw new HTTPException(403);
-  }
-
-  // TODO: Error Handling
-  const grade = await db
-    .insert(grades)
-    .values({
-      name,
-      outOf,
-      value,
-      coefficient,
-      passedAt,
-      createdAt: new Date(),
-      userId: session.user.id,
-      subjectId: subject.id,
-      periodId,
-    })
-    .returning()
-    .get();
-
-  return c.json({ grade }, 201);
-});
-
-/**
- * Get all grades
- */
-
-const getGradesQuerySchema = z.object({
-  from: z.coerce.date().optional(),
-  to: z.coerce.date().optional(),
-  limit: z.coerce.number().int().optional(),
-});
-
-app.get("/", zValidator("query", getGradesQuerySchema), async (c) => {
-  const session = c.get("session");
-  if (!session) throw new HTTPException(401);
-
-  // If email isnt verified
-  if (!session.user.emailVerified) {
-    return c.json(
-      { code: "EMAIL_NOT_VERIFIED", message: "Email verification is required" },
-      403
-    );
-  }
-
-  const { from, to, limit } = c.req.valid("query");
-
-  let allGrades = await db.query.grades.findMany({
-    where: and(
-      eq(grades.userId, session.user.id),
-      from && gte(grades.createdAt, from),
-      to && lte(grades.createdAt, to)
-    ),
-    limit: limit,
-    orderBy: desc(grades.createdAt),
+async function getMarkByIdWithSubject(markId: string) {
+  const mark = await db.query.grades.findFirst({
+    where: (marks, { eq }) => eq(marks.id, markId),
     with: {
-      subject: {
-        columns: {
-          id: true,
-          name: true,
-        },
-      },
+      subject: true,
     },
   });
-
-  allGrades = allGrades.sort(
-    (a, b) => b.passedAt.getTime() - a.passedAt.getTime()
-  );
-
-  return c.json({ grades: allGrades });
-});
+  return mark;
+}
 
 /**
  * Get a grade by ID
@@ -170,12 +54,7 @@ app.get("/:gradeId", zValidator("param", getGradeSchema), async (c) => {
 
   const { gradeId } = c.req.valid("param");
 
-  const grade = await db.query.grades.findFirst({
-    where: eq(grades.id, gradeId),
-    with: {
-      subject: true,
-    },
-  });
+  const grade = await getMarkByIdWithSubject(gradeId);
 
   if (!grade) throw new HTTPException(404);
   if (grade.userId !== session.user.id) throw new HTTPException(403);
@@ -238,6 +117,11 @@ app.patch(
     const { gradeId } = c.req.valid("param");
     const data = c.req.valid("json");
 
+    const grade = await getMarkById(gradeId);
+
+    if (!grade) throw new HTTPException(404);
+    if (grade.userId !== session.user.id) throw new HTTPException(403);
+
     // Check if the subject exists and belongs to the user
     if (data.subjectId) {
       const subject = await db.query.subjects.findFirst({
@@ -246,6 +130,7 @@ app.patch(
 
       if (!subject) throw new HTTPException(404);
       if (subject.userId !== session.user.id) throw new HTTPException(403);
+      if (subject.yearId !== grade.yearId) return c.json({ code: "SUBJECT_NOT_IN_YEAR_ERROR" }, 400);
     }
 
     // Check if the period exists and belongs to the user
@@ -256,16 +141,17 @@ app.patch(
 
       if (!period) throw new HTTPException(404);
       if (period.userId !== session.user.id) throw new HTTPException(403);
+      if (period.yearId !== grade.yearId) return c.json({ code: "PERIOD_NOT_IN_YEAR_ERROR" }, 400);
     }
 
-    const grade = await db
+    const updatedGrade = await db
       .update(grades)
       .set(data)
-      .where(and(eq(grades.id, gradeId), eq(grades.userId, session.user.id)))
+      .where(eq(grades.id, grade.id))
       .returning()
       .get();
 
-    return c.json({ grade });
+    return c.json({ grade: updatedGrade });
   }
 );
 
@@ -291,13 +177,18 @@ app.delete("/:gradeId", zValidator("param", deleteGradeSchema), async (c) => {
 
   const { gradeId } = c.req.valid("param");
 
-  const grade = await db
+  const grade = await getMarkById(gradeId);
+
+  if (!grade) throw new HTTPException(404);
+  if (grade.userId !== session.user.id) throw new HTTPException(403);
+
+  const deletedGrade = await db
     .delete(grades)
-    .where(and(eq(grades.id, gradeId), eq(grades.userId, session.user.id)))
+    .where(eq(grades.id, grade.id))
     .returning()
     .get();
 
-  return c.json({ grade });
+  return c.json({ grade: deletedGrade });
 });
 
 export default app;
