+++ apps/api/src/routes/periods.ts
@@ -16,78 +16,12 @@ const app = new Hono<{
   };
 }>();
 
-/**
- * Create a new period
- */
-
-const createPeriodSchema = z.object({
-  name: z.string().min(1).max(64),
-  startAt: z.coerce.date().optional().default(new Date()),
-  endAt: z.coerce.date().optional().default(new Date()),
-  // NEW FIELD
-  isCumulative: z.boolean().default(false),
-});
-
-app.post("/", zValidator("json", createPeriodSchema), async (c) => {
-  const session = c.get("session");
-
-  if (!session) throw new HTTPException(401);
-
-  // If email isnt verified
-  if (!session.user.emailVerified) {
-    return c.json(
-      { code: "EMAIL_NOT_VERIFIED", message: "Email verification is required" },
-      403
-    );
-  }
-
-  const { name, startAt, endAt, isCumulative } = c.req.valid("json");
-
-  if (startAt > endAt) {
-    throw new HTTPException(400);
-  }
-
-  const period = await db
-    .insert(periods)
-    .values({
-      name,
-      startAt: new Date(startAt.getTime()),
-      endAt: new Date(endAt.getTime()),
-      userId: session.user.id,
-      createdAt: new Date(),
-      isCumulative
-    })
-    .returning()
-    .get();
-
-  return c.json(period);
-});
-
-/**
- * Get all periods
- */
-app.get("/", async (c) => {
-  const session = c.get("session");
-
-  if (!session) throw new HTTPException(401);
-
-  // If email isnt verified
-  if (!session.user.emailVerified) {
-    return c.json(
-      { code: "EMAIL_NOT_VERIFIED", message: "Email verification is required" },
-      403
-    );
-  }
-
-  const allPeriods = await db.query.periods.findMany({
-    where: eq(periods.userId, session.user.id),
-    orderBy: asc(periods.startAt),
+async function getPeriodById(periodId: string) {
+  const period = await db.query.periods.findFirst({
+    where: eq(periods.id, periodId),
   });
-
-  return c.json({
-    periods: allPeriods,
-  });
-});
+  return period;
+}
 
 /**
  * Get a period by id
@@ -111,12 +45,9 @@ app.get("/:periodId", zValidator("param", getPeriodSchema), async (c) => {
   }
 
   const { periodId } = c.req.valid("param");
-
-  const period = await db.query.periods.findFirst({
-    where: and(eq(periods.id, periodId), eq(periods.userId, session.user.id)),
-  });
-
+  const period = await getPeriodById(periodId);
   if (!period) throw new HTTPException(404);
+  if (period.userId !== session.user.id) throw new HTTPException(403);
 
   return c.json(period);
 });
@@ -153,10 +84,14 @@ app.patch(
       );
     }
 
-    const { periodId } = c.req.valid("param");
     const { name, startAt, endAt, isCumulative } = c.req.valid("json");
 
-    const period = await db
+    const { periodId } = c.req.valid("param");
+    const period = await getPeriodById(periodId);
+    if (!period) throw new HTTPException(404);
+    if (period.userId !== session.user.id) throw new HTTPException(403);
+
+    const updatedPeriod = await db
       .update(periods)
       .set({
         name,
@@ -168,9 +103,7 @@ app.patch(
       .returning()
       .get();
 
-    if (!period) throw new HTTPException(404);
-
-    return c.json(period);
+    return c.json(updatedPeriod);
   }
 );
 
@@ -196,16 +129,17 @@ app.delete("/:periodId", zValidator("param", deletePeriodSchema), async (c) => {
   }
 
   const { periodId } = c.req.valid("param");
+  const period = await getPeriodById(periodId);
+  if (!period) throw new HTTPException(404);
+  if (period.userId !== session.user.id) throw new HTTPException(403);
 
-  const period = await db
+  const deletedPeriod = await db
     .delete(periods)
     .where(and(eq(periods.id, periodId), eq(periods.userId, session.user.id)))
     .returning()
     .get();
 
-  if (!period) throw new HTTPException(404);
-
-  return c.json({ period });
+  return c.json({ period: deletedPeriod });
 });
 
 export default app;
