+++ apps/api/src/routes/subjects.ts
@@ -1,8 +1,8 @@
 import { db } from "@/db";
-import { grades, periods, subjects } from "@/db/schema";
+import { periods, subjects } from "@/db/schema";
 import { type Session, type User } from "@/lib/auth";
 import { zValidator } from "@hono/zod-validator";
-import { and, asc, eq, ne, sql } from "drizzle-orm";
+import { and, asc, eq, } from "drizzle-orm";
 import { Hono } from "hono";
 import { HTTPException } from "hono/http-exception";
 import { z } from "zod";
@@ -16,211 +16,44 @@ const app = new Hono<{
   };
 }>();
 
-/**
- * Create a new subject
- */
-const createSubjectSchema = z.object({
-  name: z.string().min(1).max(64),
-  coefficient: z
-    .number()
-    .min(1)
-    .max(1000)
-    .transform((f) => Math.round(f * 100)),
-  parentId: z.string().min(1).max(64).optional().nullable(),
-  depth: z.number().int().min(0).max(1000).optional(),
-  isMainSubject: z.boolean().optional(),
-  isDisplaySubject: z.boolean().optional(),
-});
+async function getSubjectById(subjectId: string) {
+  const subject = await db.query.subjects.findFirst({
+    where: eq(subjects.id, subjectId),
+  });
+  return subject;
+}
 
-app.post("/", zValidator("json", createSubjectSchema), async (c) => {
-  const session = c.get("session");
-  if (!session) throw new HTTPException(401);
-
-  // If email isnt verified
-  if (!session.user.emailVerified) {
-    return c.json(
-      { code: "EMAIL_NOT_VERIFIED", message: "Email verification is required" },
-      403
-    );
-  }
-
-  const { name, coefficient, parentId, isMainSubject, isDisplaySubject } =
-    c.req.valid("json");
-
-  let depth = 0;
-
-  if (parentId) {
-    const parentSubject = await db.query.subjects.findFirst({
-      where: eq(subjects.id, parentId),
-    });
-
-    if (!parentSubject) throw new HTTPException(404);
-    if (parentSubject.userId !== session.user.id) throw new HTTPException(403);
-
-    depth = parentSubject.depth + 1;
-  }
-
-  // TODO: Error Handling
-  const subject = await db
-    .insert(subjects)
-    .values({
-      name,
-      coefficient,
-      parentId,
-      depth,
-      isMainSubject: isMainSubject,
-      isDisplaySubject: isDisplaySubject,
-      createdAt: new Date(),
-      userId: session.user.id,
-    })
-    .returning()
-    .get();
-
-  return c.json({ subject }, 201);
-});
-
-/**
- * Get all subjects
- */
-app.get("/", async (c) => {
-  const session = c.get("session");
-  if (!session) throw new HTTPException(401);
-
-  // If email isnt verified
-  if (!session.user.emailVerified) {
-    return c.json(
-      { code: "EMAIL_NOT_VERIFIED", message: "Email verification is required" },
-      403
-    );
-  }
-
-  const allSubjects = await db.query.subjects.findMany({
-    where: eq(subjects.userId, session.user.id),
+async function getSubjectByIdWithMarks(subjectId: string) {
+  const subject = await db.query.subjects.findFirst({
+    where: eq(subjects.id, subjectId),
     with: {
       grades: true,
     },
-    orderBy: sql`COALESCE(${subjects.parentId}, ${subjects.id}), ${subjects.parentId} IS NULL DESC, ${subjects.name} ASC`,
   });
+  return subject;
+}
 
-  return c.json({ subjects: allSubjects });
-});
-
-app.get("/organized-by-periods", async (c) => {
-  const session = c.get("session");
-  if (!session) throw new HTTPException(401);
-
-  // If email isn't verified
-  if (!session.user.emailVerified) {
-    return c.json(
-      { code: "EMAIL_NOT_VERIFIED", message: "Email verification is required" },
-      403
-    );
-  }
-
-  // 1) Fetch periods for the user
-  const userPeriods = await db.query.periods.findMany({
-    where: eq(periods.userId, session.user.id),
-    orderBy: asc(periods.startAt),
-  });
-
-  // 2) Fetch subjects with grades
-  const subjectsWithGrades = await db.query.subjects.findMany({
-    where: eq(subjects.userId, session.user.id),
+async function getSubjectByIdWithParentAndChildrenAndMarks(subjectId: string) {
+  const subject = await db.query.subjects.findFirst({
+    where: eq(subjects.id, subjectId),
     with: {
-      grades: {
-        columns: {
-          id: true,
-          name: true,
-          value: true,
-          outOf: true,
-          coefficient: true,
-          passedAt: true,
-          periodId: true,
-        },
-      },
-    },
-    orderBy: asc(subjects.name),
+      grades: true,
+      childrens: true,
+      parent: true
+    }
   });
-
-  // 3) Sort periods by start date
-  const sortedPeriods = [...userPeriods].sort(
-    (a, b) => new Date(a.startAt).getTime() - new Date(b.startAt).getTime()
-  );
-
-  // 4) Build the data structure: for each period, if isCumulative = true,
-  //    include the grades from all *previous* periods plus the current one.
-  const periodsWithSubjects = sortedPeriods.map((period, index) => {
-    // Gather relevant period IDs
-    const relevantPeriodIds = period.isCumulative
-      ? sortedPeriods.slice(0, index + 1).map((p) => p.id)
-      : [period.id];
-
-    // Now, for each subject, we include only the grades whose `periodId` is in relevantPeriodIds
-    const subjectsInPeriod = subjectsWithGrades.map((subject) => {
-      const relevantGrades = subject.grades.filter((grade) =>
-        relevantPeriodIds.includes(grade.periodId ?? "")
-      );
-
-      return {
-        ...subject,
-        grades: relevantGrades,
-      };
-    });
-
-    return {
-      period,
-      subjects: subjectsInPeriod,
-    };
-  });
-
-  // -- The following is the "full-year" logic you already have. --
-  // Calculate startAt and endAt for the "full-year" period:
-  const fullYearStartAt =
-    sortedPeriods.length > 0
-      ? new Date(sortedPeriods[0].startAt)
-      : new Date(new Date().getFullYear(), 8, 1);
-
-  const fullYearEndAt =
-    sortedPeriods.length > 0
-      ? new Date(sortedPeriods[sortedPeriods.length - 1].endAt)
-      : new Date(new Date().getFullYear() + 1, 5, 30);
-
-  // Hardcoded full-year period
-  const fullYearPeriod = {
-    id: "full-year",
-    name: "Full Year",
-    startAt: fullYearStartAt,
-    endAt: fullYearEndAt,
-    createdAt: new Date(),
-    userId: session.user.id,
-    isCumulative: false, // or true, depending on your logic
-  };
-
-  // All subjects with all grades
-  const allSubjectsWithAllGrades = subjectsWithGrades.map((subject) => ({
-    ...subject,
-    grades: subject.grades,
-  }));
-
-  // Add the full-year period to the array
-  periodsWithSubjects.push({
-    period: fullYearPeriod,
-    subjects: allSubjectsWithAllGrades,
-  });
-
-  // 5) Return the final result
-  return c.json({ periods: periodsWithSubjects });
-});
+  return subject;
+}
 
 
 // Get a specific subject organized by periods
-const getSubjectByPeriodSchema = z.object({
-  subjectId: z.string().min(1).max(64).nullable(),
-});
+// const getSubjectByPeriodSchema = z.object({
+//   subjectId: z.string().min(1).max(64).nullable(),
+// });
 
 app.get(
   "/organized-by-periods/:subjectId",
-  zValidator("param", getSubjectByPeriodSchema),
+  // zValidator("param", getSubjectByPeriodSchema),
   async (c) => {
     const session = c.get("session");
     if (!session) throw new HTTPException(401);
@@ -236,26 +69,27 @@ app.get(
       );
     }
 
-    const { subjectId } = c.req.valid("param");
+    const { subjectId } = c.req.param();
 
-    // if subjectId is null or empty, return all subjects
-    if (!subjectId) {
-      const allSubjects = await db.query.subjects.findMany({
-        where: eq(subjects.userId, session.user.id),
-        with: {
-          grades: true,
-        },
-        orderBy: sql`COALESCE(${subjects.parentId}, ${subjects.id}), ${subjects.parentId} IS NULL DESC, ${subjects.name} ASC`,
-      });
+    // useless bc subjectId never null
+    // // if subjectId is null or empty, return all subjects
+    // if (!subjectId) {
+    //   const allSubjects = await db.query.subjects.findMany({
+    //     where: eq(subjects.userId, session.user.id),
+    //     with: {
+    //       grades: true,
+    //     },
+    //     orderBy: sql`COALESCE(${subjects.parentId}, ${subjects.id}), ${subjects.parentId} IS NULL DESC, ${subjects.name} ASC`,
+    //   });
 
-      return c.json({ subjects: allSubjects });
-    }
+    //   return c.json({ subjects: allSubjects });
+    // }
 
     // Fetch the single subject + its grades
     const subject = await db.query.subjects.findFirst({
       where: and(
         eq(subjects.id, subjectId),
-        eq(subjects.userId, session.user.id)
+        // eq(subjects.userId, session.user.id)
       ),
       with: {
         grades: {
@@ -273,6 +107,7 @@ app.get(
     });
 
     if (!subject) throw new HTTPException(404);
+    if (subject.userId !== session.user.id) throw new HTTPException(403);
 
     // Fetch periods for the user and sort them
     const userPeriods = await db.query.periods.findMany({
@@ -308,11 +143,11 @@ app.get(
 /**
  * Get subject by ID
  */
-const getSubjectSchema = z.object({
-  subjectId: z.string().min(1).max(64),
-});
+// const getSubjectSchema = z.object({
+//   subjectId: z.string().min(1).max(64),
+// });
 
-app.get("/:subjectId", zValidator("param", getSubjectSchema), async (c) => {
+app.get("/:subjectId", async (c) => {
   const session = c.get("session");
   if (!session) throw new HTTPException(401);
 
@@ -327,16 +162,9 @@ app.get("/:subjectId", zValidator("param", getSubjectSchema), async (c) => {
     );
   }
 
-  const { subjectId } = c.req.valid("param");
+  const { subjectId } = c.req.param();
 
-  const subject = await db.query.subjects.findFirst({
-    where: eq(subjects.id, subjectId),
-    with: {
-      parent: true,
-      childrens: true,
-      grades: true,
-    },
-  });
+  const subject = await getSubjectByIdWithParentAndChildrenAndMarks(subjectId);
 
   if (!subject) throw new HTTPException(404);
   if (subject.userId !== session.user.id) throw new HTTPException(403);
@@ -356,6 +184,7 @@ const updateSubjectBodySchema = z.object({
     .transform((f) => Math.round(f * 100))
     .optional(),
   parentId: z.string().min(1).max(64).optional().nullable(),
+  // TODO: Dont allow user to edit the depth
   depth: z.number().int().min(0).max(1000).optional(),
   isMainSubject: z.boolean().optional(),
   isDisplaySubject: z.boolean().optional(),
@@ -388,12 +217,13 @@ app.patch(
     const data = c.req.valid("json");
 
     // Get the subject to be updated
-    const subject = await db.query.subjects.findFirst({
-      where: eq(subjects.id, subjectId),
-    });
+    const subject = await getSubjectByIdWithMarks(subjectId);
 
     if (!subject) throw new HTTPException(404);
     if (subject.userId !== session.user.id) throw new HTTPException(403);
+
+    // If subjects has mark it cannot be a category
+    if (subject.grades.length > 0 && data.isDisplaySubject) return c.json({ code: "CATEGORY_CANT_HAVE_MARKS_ERROR" }, 400);
 
     let newDepth = subject.depth;
     let depthDifference = 0;
@@ -409,13 +239,13 @@ app.patch(
         }
 
         // Get the new parent subject
-        const parentSubject = await db.query.subjects.findFirst({
-          where: eq(subjects.id, newParentId),
-        });
+        const parentSubject = await getSubjectById(newParentId);
 
         if (!parentSubject) throw new HTTPException(404);
         if (parentSubject.userId !== session.user.id)
           throw new HTTPException(403);
+        if (parentSubject.yearId !== subject.yearId)
+          return c.json({ code: "PARENT_NOT_IN_YEAR" }, 400);
 
         // Compute new depth
         newDepth = parentSubject.depth + 1;
@@ -430,12 +260,14 @@ app.patch(
       data.depth = newDepth;
     }
 
+    // TODO: Use transactions
+
     // Update the subject
     const updatedSubject = await db
       .update(subjects)
       .set(data)
       .where(
-        and(eq(subjects.id, subjectId), eq(subjects.userId, session.user.id))
+        and(eq(subjects.id, subjectId), eq(subjects.yearId, subject.yearId), eq(subjects.userId, session.user.id))
       )
       .returning()
       .get();
@@ -446,6 +278,7 @@ app.patch(
       const descendants = await db.query.subjects.findMany({
         where: and(
           eq(subjects.userId, session.user.id),
+          eq(subjects.yearId, subject.yearId),
           eq(subjects.parentId, subjectId)  // Direct children only
         ),
       });
@@ -455,6 +288,7 @@ app.patch(
         const children = await db.query.subjects.findMany({
           where: and(
             eq(subjects.userId, session.user.id),
+            eq(subjects.yearId, subject.yearId),
             eq(subjects.parentId, parentId)
           ),
         });
@@ -473,11 +307,12 @@ app.patch(
       await updateDescendantDepths(subjectId, depthDifference);
     }
 
-    // If the subject is now a category (assuming `isMainSubject === true` indicates category)
-    // delete all associated grades.
-    if (data.isDisplaySubject === true) {
-      await db.delete(grades).where(eq(grades.subjectId, subjectId));
-    }
+    // Dont silently delete user grades
+    // // If the subject is now a category (assuming `isMainSubject === true` indicates category)
+    // // delete all associated grades.
+    // if (data.isDisplaySubject === true) {
+    //   await db.delete(grades).where(eq(grades.subjectId, subjectId));
+    // }
 
     return c.json({ subject: updatedSubject });
   }
@@ -509,7 +344,12 @@ app.delete(
 
     const { subjectId } = c.req.valid("param");
 
-    const subject = await db
+    const subject = await getSubjectById(subjectId);
+
+    if (!subject) throw new HTTPException(404);
+    if (subject.userId !== session.user.id) throw new HTTPException(403);
+
+    const deletedSubject = await db
       .delete(subjects)
       .where(
         and(eq(subjects.id, subjectId), eq(subjects.userId, session.user.id))
@@ -517,7 +357,7 @@ app.delete(
       .returning()
       .get();
 
-    return c.json({ subject });
+    return c.json({ subject: deletedSubject });
   }
 );
